+++ apps/cockpit_web/lib/cockpit_web/controllers/page_controller.ex
@@ -1,0 +1,12 @@
+defmodule Tankste.CockpitWeb.PageController do
+  use Tankste.CockpitWeb, :controller
+
+  import Tankste.CockpitWeb.AuthPlug
+
+  plug :load_current_user
+  plug :require_current_user
+
+  def home(conn, _params) do
+    render(conn, :home)
+  end
+end