+++ apps/cockpit_web/lib/cockpit_web/controllers/page_html.ex
@@ -1,0 +1,5 @@
+defmodule Tankste.CockpitWeb.PageHTML do
+  use Tankste.CockpitWeb, :html
+
+  embed_templates "page_html/*"
+end