+++ apps/cockpit_web/lib/cockpit_web/controllers/station_html.ex
@@ -1,0 +1,5 @@
+defmodule Tankste.CockpitWeb.StationHTML do
+  use Tankste.CockpitWeb, :html
+
+  embed_templates "station_html/*"
+end