+++ apps/sponsor_web/lib/sponsor_web.ex
@@ -38,8 +38,6 @@ defmodule Tankste.SponsorWeb do
 
   defp view_helpers do
     quote do
-      use Phoenix.HTML
-
       import Phoenix.View
 
       alias Tankste.SponsorWeb.Router.Helpers, as: Routes